chore: organize node validators. Move isExtendedCommandClassDeclaration to it's own file
