+++ migration/src/index.ts
@@ -6,10 +6,11 @@ import {ESLint} from 'eslint'
 
 import {createClassElementsFromModuleExports} from './createClassElementFromModuleExports.js'
 import {createCommandClass} from './createCommandClass.js'
-import {isExtendedCommandClassDeclaration, isModuleExports} from './isModuleExports.js'
-import {isRunFunctionDecl} from './isRunFunctionDecl.js'
+import {isModuleExports} from './node-validators/isModuleExports'
+import {isRunFunctionDecl} from './node-validators/isRunFunctionDecl'
 import {nullTransformationContext} from './nullTransformationContext.js'
-import {isMigrationCandidate} from './isMigrationCandidate.js'
+import {isMigrationCandidate} from './node-validators/isMigrationCandidate'
+import {isExtendedCommandClassDeclaration} from './node-validators/isExtendedCommandClassDeclaration'
 
 const commonImports = `import {createRequire} from 'node:module'
 import color from '@heroku-cli/color'
