+++ migration/src/node-validators/isModuleExports.ts
@@ -22,11 +22,3 @@ export function isModuleExports(node: ts.Node): node is (ts.Node & {right: ts.Ob
         ts.isObjectLiteralExpression(node.right)
 }
 
-export const isExtendedCommandClassDeclaration = (node: ts.Node): node is ts.ClassDeclaration => {
-  return ts.isClassDeclaration(node) &&
-    node.heritageClauses.some(
-      clause => clause.types.some(type => (
-        ts.isIdentifier(type.expression) && type.expression.escapedText === 'Command'),
-      ),
-    )
-}
